+++ crates/mds-tui/src/message.rs
@@ -35,4 +35,5 @@ pub enum Popup {
     SearchBox,
     ErrorBox,
     IpInfoPopUp,
+    Keybindings,
 }